intersection-observer/remove-element.html

<!DOCTYPE html>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="./resources/intersection-observer-test-utils.js"></script>

<style>
pre, #log {
  position: absolute;
  top: 0;
  left: 200px;
}
#root {
  display: inline-block;
  overflow-y: scroll;
  height: 200px;
  border: 3px solid black;
}
#target {
  width: 100px;
  height: 100px;
  background-color: green;
}
.spacer {
  height: 300px;
}
</style>

<div id="root">
  <div id="leading-space" class="spacer"></div>
  <div id="target"></div>
  <div id="trailing-space" class="spacer"</div>
</div>

<script>
var entries = [];
var root, target, trailingSpace;

runTestCycle(function() {
  assert_equals(window.innerWidth, 800, "Window must be 800 pixels wide.");
  assert_equals(window.innerHeight, 600, "Window must be 600 pixels high.");

  target = document.getElementById("target");
  assert_true(!!target, "Target exists");
  trailingSpace = document.getElementById("trailing-space");
  assert_true(!!trailingSpace, "TrailingSpace exists");
  root = document.getElementById("root");
  assert_true(!!root, "Root exists");
  var observer = new IntersectionObserver(function(changes) {
    entries = entries.concat(changes)
  }, {root: root});
  observer.observe(target);
  entries = entries.concat(observer.takeRecords());
  assert_equals(entries.length, 0, "No initial notifications.");
  runTestCycle(step0, "First rAF");
}, "Verify that not-intersecting notifications are sent when a target is removed from the DOM tree.");

function step0() {
  root.scrollTop = 150;
  runTestCycle(step1, "root.scrollTop = 150");
  assert_equals(entries.length, 0, "No notifications after first rAF.");
}

function step1() {
  root.removeChild(target);
  runTestCycle(step2, "root.removeChild(target).");
  checkEntry(entries, 0, [11, 111, 161, 261, 11, 111, 161, 211, 11, 111, 11, 211, target]);
}

function step2() {
  root.scrollTop = 0;
  root.insertBefore(target, trailingSpace);
  runTestCycle(step3, "root.insertBefore(target, trailingSpace).");
  checkEntry(entries, 1, [0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, target]);
}

function step3() {
  root.scrollTop = 150;
  runTestCycle(step4, "root.scrollTop = 150 after reinserting target.");
  checkEntry(entries, 1);
}

function step4() {
  checkEntry(entries, 2, [11, 111, 161, 261, 11, 111, 161, 211, 11, 111, 11, 211, target]);
}
</script>
